Move world to ARB program
The GLSL C interface sucks.  Because we want to move the entire renderer to shaders, and take advantage of higher-quality and faster paths for e.g. sky and water, we use ARB programs instead.  For a Quake-style renderer, the only real missing features in ARB programs are texGrad instructions (which would have been nice to have) and texture arrays (not as big a deal).